Set up new classes to recognise acoustic types
Issue #252 Set up separate class for determining amount of each type of acoustic content. Add new class to hold visualization methods. Having difficulty with writing strings to image so play around with ImageTrack.cs but cannot fix problem.
Set up BaseContent.cs class with eventual intention to do inheritance.
