+++ src/AudioAnalysisTools/ContentDescriptionTools/ContentDescription.cs
@@ -12,6 +12,7 @@ namespace AudioAnalysisTools.ContentDescriptionTools
     using System.Linq;
     using Acoustics.Shared;
     using Acoustics.Shared.Csv;
+    using AudioAnalysisTools.ContentDescriptionTools.ContentTypes;
     using AudioAnalysisTools.DSP;
     using AudioAnalysisTools.LongDurationSpectrograms;
     using AudioAnalysisTools.StandardSpectrograms;
@@ -172,10 +173,13 @@ namespace AudioAnalysisTools.ContentDescriptionTools
                 var descriptionResult = new DescriptionResult(elapsedMinutes + i);
 
                 // now send indices to various content searches
-                descriptionResult.AddDescription(WindContent.GetStrongWindContent(oneMinuteOfIndices));
-                descriptionResult.AddDescription(WindContent.GetLightWindContent(oneMinuteOfIndices));
-                descriptionResult.AddDescription(RainContent.GetStrongRainContent(oneMinuteOfIndices));
-                descriptionResult.AddDescription(RainContent.GetLightRainContent(oneMinuteOfIndices));
+                descriptionResult.AddDescription(WindStrong1.GetContent(oneMinuteOfIndices));
+                descriptionResult.AddDescription(WindLight1.GetContent(oneMinuteOfIndices));
+                descriptionResult.AddDescription(RainLight1.GetContent(oneMinuteOfIndices));
+
+                // yet to do following
+                //descriptionResult.AddDescription(RainHeavy1.GetContent(oneMinuteOfIndices));
+                //descriptionResult.AddDescription(RainHeavy2.GetContent(oneMinuteOfIndices));
 
                 results.Add(descriptionResult);
             }
@@ -183,8 +187,7 @@ namespace AudioAnalysisTools.ContentDescriptionTools
             return results;
         }
 
-        public static Dictionary<string, double[]> GetIndicesForOneMinute(Dictionary<string, double[,]> allIndices,
-            int rowId)
+        public static Dictionary<string, double[]> GetIndicesForOneMinute(Dictionary<string, double[,]> allIndices, int rowId)
         {
             var opIndices = new Dictionary<string, double[]>();
 
@@ -259,46 +262,6 @@ namespace AudioAnalysisTools.ContentDescriptionTools
             return list.ToArray();
         }
 
-        public static void DrawNormalisedIndexMatrices(DirectoryInfo dir, string baseName, Dictionary<string, double[,]> dictionary)
-        {
-            var list = new List<Image>();
-            foreach (string key in IndexNames)
-            {
-                var bmp = ImageTools.DrawReversedMatrixWithoutNormalisation(dictionary[key]);
-
-                // need to rotate spectrogram to get correct orientation.
-                bmp.RotateFlip(RotateFlipType.Rotate270FlipNone);
-
-                // draw grid lines and add axis scales
-                var xAxisPixelDuration = TimeSpan.FromSeconds(60);
-                var fullDuration = TimeSpan.FromTicks(xAxisPixelDuration.Ticks * bmp.Width);
-                var freqScale = new FrequencyScale(11025, 512, 1000);
-                SpectrogramTools.DrawGridLinesOnImage((Bitmap)bmp, TimeSpan.Zero, fullDuration, xAxisPixelDuration, freqScale);
-                const int trackHeight = 20;
-                var recordingStartDate = default(DateTimeOffset);
-                var timeBmp = ImageTrack.DrawTimeTrack(fullDuration, recordingStartDate, bmp.Width, trackHeight);
-                var array = new Image[2];
-                array[0] = bmp;
-                array[1] = timeBmp;
-                var image = ImageTools.CombineImagesVertically(array);
-
-                // add a header to the spectrogram
-                var header = new Bitmap(image.Width, 20);
-                Graphics g = Graphics.FromImage(header);
-                g.Clear(Color.LightGray);
-                g.SmoothingMode = SmoothingMode.AntiAlias;
-                g.InterpolationMode = InterpolationMode.HighQualityBicubic;
-                g.PixelOffsetMode = PixelOffsetMode.HighQuality;
-                g.DrawString(key, new Font("Tahoma", 9), Brushes.Black, 4, 4);
-                list.Add(ImageTools.CombineImagesVertically(new List<Image>(new[] { header, image })));
-            }
-
-            // save the image - the directory for the path must exist
-            var path = Path.Combine(dir.FullName, baseName + "__Towsey.Acoustic.GreyScaleImages.png");
-            var indexImage = ImageTools.CombineImagesInLine(list);
-            indexImage?.Save(path);
-        }
-
         public static Dictionary<string, Plot> ConvertResultsToPlots(List<DescriptionResult> results, int plotLength, int plotStart)
         {
             var plots = new Dictionary<string, Plot>();
@@ -370,34 +333,6 @@ namespace AudioAnalysisTools.ContentDescriptionTools
             }
 
             return list;
-        }
-
-        public static Image DrawLdfcSpectrogramWithContentScoreTracks(Image ldfcSpectrogram, List<Plot> contentScores)
-        {
-            int trackHeight = 30;
-            int width = ldfcSpectrogram.Width;
-            var imageList = new List<Image>
-            {
-                ldfcSpectrogram,
-            };
-
-            foreach (var plot in contentScores)
-            {
-                var track = new ImageTrack(TrackType.scoreArrayNamed, plot.data)
-                {
-                    Name = plot.title,
-                    ScoreMin = 0.0, // plot.data.Min(),
-                    ScoreMax = 1.0, // plot.data.Max(),
-                    Height = trackHeight,
-                    topOffset = 0,
-                    ScoreThreshold = plot.threshold,
-                };
-                var bmp = new Bitmap(width, trackHeight);
-                imageList.Add(track.DrawNamedScoreArrayTrack(bmp));
-            }
-
-            Image bmp3 = ImageTools.CombineImagesVertically(imageList);
-            return bmp3;
         }
 
         /// <summary>
